♻️ recycle: 컴포넌트 분리 및 상수화 진행
