+++ src/containers/FeedContainer.jsx
@@ -4,11 +4,12 @@ import styled from '@emotion/styled';
 
 import { loadImage } from '../services/api';
 
-import { MdFace } from 'react-icons/md';
-import { BsThreeDots, BsBookmark } from 'react-icons/bs';
-import { AiOutlineHeart } from 'react-icons/ai';
-import { FaRegCommentDots, FaRegSmile } from 'react-icons/fa';
-import { GrSend } from 'react-icons/gr';
+import FeedImage from '../components/feed/FeedImage';
+import FeedHeader from '../components/feed/FeedHeader';
+import FeedIcons from '../components/feed/FeedIcons';
+import ThumbsUp from '../components/feed/ThumbsUp';
+import Comments from '../components/feed/Comments';
+import CommentWriteField from '../components/feed/CommentWriteField';
 
 export default function FeedContainer({ feed }) {
   const { username, img, thumbs_up, comments } = feed;
@@ -41,7 +42,7 @@ export default function FeedContainer({ feed }) {
     }
   };
 
-  const handleClick = () => {
+  const handleClickPostingButton = () => {
     if (commentRef.current.value !== '') {
       addNewComment();
     }
@@ -49,56 +50,21 @@ export default function FeedContainer({ feed }) {
 
   return (
     <FeedBox imageLoading={imageLoading}>
-      <Header>
-        <UserField>
-          <MdFace size={18} />
-          {username}
-        </UserField>
-        <BsThreeDots />
-      </Header>
-      <Image>
-        <img
-          src={imageSrc}
-          alt="instagram-image"
-          onLoad={() => setImageloading(false)}
-        />
-      </Image>
-      <ToolBar>
-        <div style={{ display: 'flex', gap: '8px', fontSize: '18px' }}>
-          <AiOutlineHeart />
-          <FaRegCommentDots />
-          <GrSend />
-        </div>
-        <div>
-          <BsBookmark />
-        </div>
-      </ToolBar>
-      <ThumbsUp>좋아요 {thumbs_up}개</ThumbsUp>
-      <ul style={{ marginBottom: '8px' }}>
-        {commentList.map((item, index) => (
-          <CommentField key={index}>
-            <CommentUsername>{item.name}</CommentUsername>
-            <p>{item.comment}</p>
-          </CommentField>
-        ))}
-      </ul>
-      <CommentWriteField>
-        <FaRegSmile size={18} />
-        <InputField
-          placeholder="댓글달기..."
-          onKeyPress={handleKeyPress}
-          ref={commentRef}
-        />
-        <PostingButton type="button" onClick={handleClick}>
-          게시
-        </PostingButton>
-      </CommentWriteField>
+      <FeedHeader username={username} />
+      <FeedImage imageSrc={imageSrc} setImageSrc={setImageSrc} />
+      <FeedIcons />
+      <ThumbsUp thumbs_up={thumbs_up} />
+      <Comments commentList={commentList} />
+      <CommentWriteField
+        handleKeyPress={handleKeyPress}
+        commentRef={commentRef}
+        onClickPostingButton={handleClickPostingButton}
+      />
     </FeedBox>
   );
 }
 
 const FeedBox = styled.div(
-  /* image가 로딩된 이후에(imageLoading === false) 피드가 보입니다. */
   (props) => ({
     display: props.imageLoading ? 'none' : 'block',
   }),
@@ -110,67 +76,3 @@ const FeedBox = styled.div(
     border: '1px solid lightgray',
   }
 );
-
-const Header = styled.div({
-  padding: '8px',
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'space-between',
-});
-
-const UserField = styled.div({
-  gap: '8px',
-  display: 'flex',
-  fontWeight: 'bold',
-  alignItems: 'center',
-});
-
-const Image = styled.div({
-  maxHeight: '500px',
-  overflow: 'hidden',
-  '& img': {
-    width: '100%',
-  },
-});
-
-const ToolBar = styled.div({
-  display: 'flex',
-  padding: '8px',
-  justifyContent: 'space-between',
-});
-
-const ThumbsUp = styled.div({
-  padding: '8px',
-  fontSize: '12px',
-  fontWeight: 'bold',
-});
-
-const CommentField = styled.li({
-  display: 'flex',
-  padding: '3px 8px',
-  fontSize: '12px',
-});
-
-const CommentUsername = styled.div({
-  fontWeight: 'bold',
-  marginRight: '5px',
-});
-
-const CommentWriteField = styled.div({
-  display: 'flex',
-  alignItems: 'center',
-  borderTop: '1px solid #F1F1F1',
-  padding: '12px 8px',
-});
-
-const InputField = styled.input({
-  flex: '1',
-  padding: '4px',
-});
-
-const PostingButton = styled.button({
-  color: '#0095F6',
-  fontWeight: 'bold',
-  cursor: 'pointer',
-  background: 'transparent',
-});
